Calculate the axon lengths in annotated brain regions

Calculate the axon lengths in annotated brain regions

snp_calculateProjections(
  PathOfNeuron,
  Hemisphere,
  StructureIDOfTargets,
  Annotation,
  Resolution,
  NumOfCores,
  OptionMetric = "LengthOfAxon"
)

Arguments

PathOfNeuron

Full path of SWC file for each neuron

Hemisphere

Which hemisphere does the neuron reside in. Note that neurons in right hemisphere were flipped to the left side.

StructureIDOfTargets

Structure ID of brain regions or domains in the annotation. Column name of returned matrix is the same as StuctureIDOfTargets.

Annotation

3D array with each element assigned a positive value

Resolution

Voxel size of each annotation, e.g., 100, 50, 25, 10 um

NumOfCores

Number of cores for parallel computation

OptionMetric

Default value is "LengthOfAxon". ("LengthOfAxon", "NumberOfEndNodes", "NumberOfBranchNodes")

Value

projection matrix

Details

Annotated brain regions can be any 3D array with an unique index denoting a brain region or domain